Improve pkg struct: Move all comicinfo data into `dataprovider` pkg
