+++ internal/dataprovider/scanner/parse.go
