+++ internal/dataprovider/scanner/parse_test.go
