+++ internal/dataprovider/scanner/scanner.go
